Make the game continue if tail moves out the way
It used to end if the tail was where newHead was, even if it would
move out of the way at the point.

This change also meant that the game.makeFood and randomCoord now
would take arguments as to which coordinates should NOT be
considered.

That's probably good for the future...